switched to UEFI tree rather than recreating tree every time
