+++ edk2toolext/tests/minimal_uefi_tree/Conf/build_rule.template
@@ -1,0 +1,1 @@
+hello
