+++ edk2toolext/tests/minimal_uefi_tree/Conf/target.template
@@ -1,0 +1,4 @@
+ACTIVE_PLATFORM = Test.dsc
+TOOL_CHAIN_TAG = test
+TARGET_ARCH = X64
+TARGET = DEBUG
