+++ edk2toolext/tests/minimal_uefi_tree/Conf/tools_def.template
@@ -1,0 +1,1 @@
+hello
