+++ edk2toolext/tests/minimal_uefi_tree/Test.dsc
@@ -1,0 +1,2 @@
+[Defines]
+  OUTPUT_DIRECTORY = Build
