+++ edk2toolext/tests/test_edk2_setup.py
@@ -7,44 +7,11 @@
 ##
 import unittest
 from edk2toolext.invocables.edk2_setup import Edk2PlatformSetup
-import tempfile
 import sys
 import os
 import logging
 from importlib import reload
 from edk2toolext.environment import shell_environment
-
-class_info = '''
-from edk2toolext.invocables.edk2_setup import SetupSettingsManager
-class TestSettingsManager(SetupSettingsManager):
-
-    def GetActiveScopes(self):
-        return []
-
-    def GetWorkspaceRoot(self):
-        return WORKSPACE
-
-    def GetRequiredSubmodules(self):
-        return []
-
-    def AddCommandLineOptions(self, parserObj):
-        pass
-
-    def RetrieveCommandLineOptions(self, args):
-        pass
-
-    def GetName(self):
-        return "TestSetup"
-
-    def GetArchitecturesSupported(self):
-        return []
-
-    def GetPackagesSupported(self):
-        return []
-
-    def GetTargetsSupported(self):
-        return []
-'''
 
 
 class TestEdk2Setup(unittest.TestCase):
@@ -75,21 +42,9 @@ class TestEdk2Setup(unittest.TestCase):
         builder = Edk2PlatformSetup()
         self.assertIsNotNone(builder)
 
-    @classmethod
-    def write_to_file(cls, path, contents, close=True):
-        f = open(path, "w")
-        f.writelines(contents)
-        if close:
-            f.close()
-
     def test_ci_setup(self):
-        WORKSPACE = tempfile.mkdtemp()
         builder = Edk2PlatformSetup()
-        workspace_abs = os.path.abspath(WORKSPACE).replace("\\", "\\\\")
-        settings_file = os.path.abspath(os.path.join(WORKSPACE, "settings.py"))
-        real_class = class_info.replace("WORKSPACE", f'"{workspace_abs}"')
-        print(settings_file)
-        TestEdk2Setup.write_to_file(settings_file, real_class)
+        settings_file = os.path.join(os.path.dirname(__file__), "minimal_uefi_tree", "PlatformBuild.py")
         sys.argv = ["stuart_setup", "-c", settings_file]
         try:
             builder.Invoke()
